separate ensure and metadata
this PR makes it possible to get the metadata of a topic without
creating it, which is causing some bugs in nucleus.  it makes it so
that ensure_topic needs to be called before topic creation, and
metadata will not create a topic when called.

metadata now filters out non-existent topics in its return, so if the
topic metadata is missing from the reply, the topic does not exist.